cleanup deprecated functions
