+++ includes/template-functions.php
@@ -1,7 +1,5 @@
 <?php
 
-use Directorist\Helper;
-
 if ( ! defined( 'ABSPATH' ) ) {
     exit;
 }
@@ -105,45 +103,6 @@ function atbdp_get_widget_template( $template, $args = [] ) {
 
 function atbdp_get_widget_template_path( $template ) {
     $template = 'widgets/' . $template;
-
-    return atbdp_get_template_path( $template );
-}
-
-function atbdp_get_shortcode_template_paths( $template_file ) {
-    _deprecated_function( __FUNCTION__, '7.0', 'Helper::get_template()' );
-    $theme_template_file  = '/directorist/shortcodes/' . $template_file . '.php';
-    $theme_template_path  = get_stylesheet_directory() . $theme_template_file;
-    $plugin_template_path = Helper::template_directory() . 'public-templates/shortcodes/' . $template_file . '.php';
-
-    return [
-        'theme'  => $theme_template_path,
-        'plugin' => $plugin_template_path,
-    ];
-}
-
-function atbdp_get_shortcode_template( $template, $args = [] ) {
-    _deprecated_function( __FUNCTION__, '7.0', 'Helper::get_template()' );
-
-    return Helper::get_template( $template, $args );
-}
-
-function atbdp_return_shortcode_template( $template, $args = [] ) {
-    _deprecated_function( __FUNCTION__, '7.0', 'Helper::get_template_contents()' );
-
-    return Helper::get_template_contents( $template, $args );
-}
-
-function atbdp_return_widget_template( $template, $args = [] ) {
-    _deprecated_function( __FUNCTION__, '7.0' );
-    ob_start();
-    atbdp_get_widget_template( $template, $args );
-
-    return ob_get_clean();
-}
-
-function atbdp_get_shortcode_template_path( $template ) {
-    _deprecated_function( __FUNCTION__, '7.0' );
-    $template = 'shortcodes/' . $template;
 
     return atbdp_get_template_path( $template );
 }
